requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
distlib==0.3.7
Django==4.1.6
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-recaptcha==3.0.0
docopt==0.6.2
filelock==3.12.3
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai==0.26.5
Pillow==9.4.0
pipreqs==0.4.11
platformdirs==3.10.0
pypng==0.20220715.0
qrcode==7.4.2
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tqdm==4.64.1
typing_extensions==4.8.0
urllib3==1.26.14
virtualenv==20.24.3
yarg==0.1.9
yarl==1.8.2
